respond to review comments
